+++ src/component.rs
@@ -86,6 +86,11 @@ where
         self.sender.send(msg)
     }
 
+    /// Get a sender to send messages to this component.
+    pub fn sender(&self) -> Sender<Model::Msg> {
+        self.sender.clone()
+    }
+
     /// Returns the root widget of this component's widgets.
     /// Can be used by the parent [`Widgets::connect_components`](fn@crate::Widgets::connect_components) to connect the root widget
     /// to the parent's widgets.